requirements.txt
fastapi
uvicorn
python-multipart
pydantic
pytesseract
Pillow
PyMuPDF
python-dotenv
mysql-connector-python
a2wsgi
jinja2
qrcode
numpy

